pyproject.toml
[tool.poetry]
name = "hebi"
version = "0.1.0"
description = "generic process bridge for discord"
authors = ["Sofia \"dzshn\" N. L. <zshn@pm.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.9"
hikari_shim = "^1.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
